+++ Librarys/GRN.h
@@ -5,5 +5,8 @@ int getRandomNumber(int min, int max);
 
 int* getArrayOfRandomNumbers(int min, int max, int count);
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <time.h>
 #include "GRN.c"
 #endif // !GENERATE_RANDOM_NUMBER_NUM
